added add evaluator button
