+++ __tests__/pages/admin/evaluation-administrations/[id]/evaluees/[evaluation_result_id]/evaluators/evaluators.spec.ts
@@ -185,6 +185,8 @@ test.describe("Admin - Select Evaluators", () => {
       await expect(page.getByText("User, Sample")).toBeVisible()
       await expect(page.getByText("pending")).toBeVisible()
 
+      await expect(page.getByRole("link", { name: "Add Evaluator" })).toBeVisible()
+
       await expect(page.getByRole("link", { name: "PM Evaluation" })).toBeVisible()
       await expect(page.getByRole("link", { name: "Peer Evaluation" })).toBeVisible()
       await expect(page.getByRole("link", { name: "Code Reviewer Evaluation" })).toBeVisible()
