+++ src/pages/admin/evaluation-administrations/[id]/evaluees/[evaluation_result_id]/evaluators/evaluators.tsx
@@ -1,7 +1,7 @@
 import { Divider } from "../../../../../../../components/ui/divider/divider"
 import { EvaluatorsRoles } from "../../../../../../../features/admin/evaluation-administrations/[id]/evaluees/[evaluation_result_id]/evaluators/evaluators-roles"
 import { EvaluatorsList } from "../../../../../../../features/admin/evaluation-administrations/[id]/evaluees/[evaluation_result_id]/evaluators/evaluators-list"
-import { EvaluatorsUser } from "../../../../../../../features/admin/evaluation-administrations/[id]/evaluees/[evaluation_result_id]/evaluators/EvaluatorsUser"
+import { EvaluatorsUser } from "../../../../../../../features/admin/evaluation-administrations/[id]/evaluees/[evaluation_result_id]/evaluators/evaluators-user"
 
 export default function Evaluators() {
   return (
